Initial Commit with final oe to play 2 players Tic Tac Toe Game
I recommend that you check out the tic-tac-toe game before continuing with the game. One of the features that you’ll notice is that there is a numbered list to the right of the game’s board. This list gives you a history of all of the moves that have occurred in the game, and it is updated as the game progresses.